feat(core): close legacy core loop TODOs
- Integrate Phantom Flash counter damage into DamagePipeline
- Implement dynamic PlayerName and PlayerPlaytime components
- Update SaveManager to extract real player data for save headers
- Add comprehensive integration tests for counter pipeline and save semantics
- Archive track todo-closure-core-loop_20260208

+++ src/game/data/SaveData.hpp
@@ -21,7 +21,7 @@ struct SaveHeader {
   std::string name;
   std::string characterClass;
   int level = 1;
-  double playtime = 0.0;
+  int64_t playtime = 0;
   int64_t timestamp = 0;
   uint32_t version = 1;
